Separate data generation from data initialization and create npm seeding script
