+++ package.json
@@ -7,7 +7,9 @@
   "engines": {
     "node": ">=6.13.0"
   },
-  "scripts": {},
+  "scripts": {
+    "seed": "node ./data/sampleData/dataInitializer.js"
+  },
   "dependencies": {
     "faker": "^4.1.0",
     "mongodb": "^3.1.1",
